초기데이터 세팅 클래스 추가
